Changed DuotoneProcessor into class and that extends Processor.  Use processor name passed as an argument
